Use bare domain for API binding link
The link is broken on www.

The reason it links to www is a holdover from issue #1400 six weeks ago.
At the time, https://humbughq.com used TLS SNI and downloading via wget
gave a certificate error, so we changed it to www as a workaround.
However, https://zulip.com does not use SNI so it's safe to revert this
workaround.

(imported from commit c87a6eb3ceab92c377edb976216be615625fafce)