Fix ~/.gitattributes syntax.
For git that's not between version 1.8.1.1 and 1.8.1.6, the previous
version did not actually exclude those things from the tarballs on
some git versions, due to changes in the syntax for that file that
were made in 1.8.1.1 and reverted in 1.8.1.6.

Kevin's also adding a check that we're not including any of the files
intended to be excluded here.

(imported from commit 503ec24ed8db4cc32c29bdf5a361e5013a02b995)