Fix OSX settings-gear-vs.-scrollbar annoyance.
On OSX Chrome, if you scrolled and then went to the settings gear
quickly, the scrollbar could stick and obscure the triangle part
of the gear menu, which was annoying.  (You don't actually have
to click the triangle to pull up the menu, but that's the
affordance.)

This change adds a little margin to the right of the gear.

(imported from commit fa64122d913a17b765d00802184009eaaeaef0b1)